requirements.txt
fastapi
fastapi-jwt-auth
motor~=3.0.0
pymongo~=4.1.1


PyYAML~=6.0

Jinja2~=3.1.2
watchdog~=2.1.8

humanize~=4.1.0
openxmllib-py3~=2.0.0

pyzmq~=23.0.0
moviepy~=1.0.3
ipython ~=7.33.0
matplotlib-inline~=0.1.3
opencv-python~=4.5.5.64

pdf2image~=1.16.0

PyMuPDF~=1.19.6
matplotlib~=3.5.2
ocrmypdf~=13.4.5
Cython~=3.0a1

wfastcgi~=3.0.0.0


elasticsearch~=6.8.2
# fastapi~=0.78.0
gunicorn~=20.1.0
uvicorn[standard]
a2wsgi~=1.4.1
python-jose~=3.3.0
passlib~=1.7.4
python-multipart~=0.0.5
pdfminer.six
pdfplumber~=0.7.1
bcrypt
kink~=0.6.5
hypercorn[trio]
icoextract~=0.1.4

pillow-avif-plugin~=1.2.2
tika~=1.24
# sentence_transformers~=2.2.2
# pyvi~=0.1.1
# pandas~=1.5.1
psutil~=5.9.3

module-dependencies~=0.2.5